ref: move types to file
